+++ ribs/archives/__init__.py
@@ -34,6 +34,7 @@ This subpackage also contains several utilities associated with the archives.
     ArchiveStats
     cqd_score
     CQDScoreResult
+    k_means_centroids
 """
 
 from ribs.archives._add_status import AddStatus
@@ -43,7 +44,7 @@ from ribs.archives._archive_stats import ArchiveStats
 from ribs.archives._array_store import ArrayStore
 from ribs.archives._categorical_archive import CategoricalArchive
 from ribs.archives._cqd_score import CQDScoreResult, cqd_score
-from ribs.archives._cvt_archive import CVTArchive
+from ribs.archives._cvt_archive import CVTArchive, k_means_centroids
 from ribs.archives._density_archive import DensityArchive
 from ribs.archives._dns_archive import DNSArchive
 from ribs.archives._grid_archive import GridArchive
@@ -51,6 +52,7 @@ from ribs.archives._proximity_archive import ProximityArchive
 from ribs.archives._sliding_boundaries_archive import SlidingBoundariesArchive
 
 __all__ = [
+    # Archives.
     "CategoricalArchive",
     "CVTArchive",
     "DensityArchive",
@@ -59,10 +61,12 @@ __all__ = [
     "ProximityArchive",
     "SlidingBoundariesArchive",
     "ArchiveBase",
+    # Utilities.
     "ArrayStore",
     "AddStatus",
     "ArchiveDataFrame",
     "ArchiveStats",
-    "CQDScoreResult",
     "cqd_score",
+    "CQDScoreResult",
+    "k_means_centroids",
 ]